Run a separate Python process just for reporting shared metrics
My initial plan to have a dedicated thread in L1b worker 0 that
servers Prometheus metrics using the method `start_http_server`
in `prometheus_client` package did not work because L1b seems
to block the entire Python process while it's waiting for a
callback from the L1 process. Instead, we can run this thread
in its own process, so it's always providing the current metrics
from the shared directory.